chore: added the ability to migrate different types
