+++ scripts/convert-docs.js
@@ -1,47 +1,21 @@
 const fs = require('fs');
-const path = require('path');
-const TurndownService = require('turndown');
+const { TYPES } = require('./constants');
+const toMarkdown = require('./converters/to-markdown');
 
-const getCategories = require('./utils/get-categories');
-const {
-  BASE_DIR,
-  GATSBY_CONTENT_TYPES,
-  GATSBY_TEMPLATE,
-} = require('./constants');
-
-const getFrontmatter = (type, doc) => `---
-title: ${doc.title.replace(':', '-')}
-contentType: ${GATSBY_CONTENT_TYPES[type]}
-template: ${GATSBY_TEMPLATE[type]}
----
-
-`;
+// TODO: no magic string
+const converters = {
+  page: toMarkdown,
+  api_doc: toMarkdown,
+  release_notes: toMarkdown,
+  release_notes_platform: toMarkdown,
+  troubleshooting_doc: toMarkdown,
+  nr1_announcement: toMarkdown,
+  attribute_definition: toMarkdown,
+};
 
 const convertDocs = (docs) => {
   docs.flat().forEach((doc) => {
-    const dir = path.join(BASE_DIR, ...getCategories(doc.docUrl));
-    const slug = doc.docUrl.split('/').slice(-1);
-    const fileName = `${dir}/${slug}.mdx`;
-
-    // Create frontmatter based on content type
-    const frontmatter = getFrontmatter(doc.type, doc);
-
-    // Convert content to markdown
-    const turndownService = new TurndownService({
-      headingStyle: 'atx',
-    });
-
-    turndownService.addRule('codeBlocks', {
-      filter: ['pre'],
-      replacement: function (content) {
-        return `~~~\n${content}\n~~~\n`;
-      },
-    });
-
-    turndownService.keep(['table']);
-
-    const bodyContent = doc.body ? turndownService.turndown(doc.body) : '';
-    const content = frontmatter + bodyContent;
+    const { content, fileName } = converters[doc.type](doc);
 
     // Write the file
     fs.writeFile(fileName, content, (err) => {
